+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Use a default value of 60 seconds if a `Retry-After` header is not present. ([#1443](https://github.com/getsentry/sentry-dotnet/issues/1443))
+
 ## 3.15.0
 
 ### Features